Updated scripting API
Updated file api
Updated binaries

+++ src/api/fs.rs
@@ -9,8 +9,14 @@ use folderio::FolderIO;
 
 pub const API_KEY:&str = "api";
 pub const GALE_KEY:&str = "_gale";
+pub const WEB_ROOT_KEY:&str = "wr";
+pub const SERVER_ROOT_KEY:&str = "sr";
 pub const DATA_ROOT_KEY:&str = "dr"; 
 pub const FILE_API:&str = "fs";
+
+pub const DATA_SPACE:u8 = 0;
+pub const SERVER_SPACE:u8 = 1;
+pub const WEB_SPACE:u8 = 2;
 
 pub fn load(engine:&Ducc) -> bool{
     let api_res:Result<Value, _> = engine.globals().get(API_KEY);
@@ -18,6 +24,10 @@ pub fn load(engine:&Ducc) -> bool{
     let api = api_obj.as_object().unwrap();
 
     let fileio = engine.create_object();
+    fileio.set("DATA_SPACE", DATA_SPACE).unwrap();
+    fileio.set("SERVER_SPACE", SERVER_SPACE).unwrap();
+    fileio.set("WEB_SPACE", WEB_SPACE).unwrap();
+
     fileio.set("create", engine.create_function(file_create)).unwrap();
     fileio.set("writeText", engine.create_function(file_write_text)).unwrap();
     fileio.set("readText", engine.create_function(file_read_text)).unwrap();
@@ -38,24 +48,32 @@ pub fn load(engine:&Ducc) -> bool{
     return true;
 }
 
+fn get_base_dir(engine:&Ducc, space:u8) -> String{
+    let robj:Object = engine.globals().get(GALE_KEY).unwrap();
+    match space{
+        DATA_SPACE => robj.get(DATA_ROOT_KEY).unwrap(),
+        SERVER_SPACE => robj.get(SERVER_ROOT_KEY).unwrap(),
+        WEB_SPACE => robj.get(WEB_ROOT_KEY).unwrap(),
+        _=> robj.get(DATA_ROOT_KEY).unwrap()
+    }
+}
 
 pub fn file_create(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FileIO::create(fpath)));
         }else{
-            error!("Invalid argument, expected string");
+            error!("Invalid argument, expected number, string");
             return Ok(Value::Boolean(false));    
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -63,24 +81,22 @@ pub fn file_create(inv: Invocation) -> Result<Value, DuccError>{
 pub fn file_write_text(inv: Invocation) -> Result<Value, DuccError> {
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 2{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-        
-        let fpath_res = args.get(0);
-        let text_res = args.get(1);
-        
-        if fpath_res.is_string() && text_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 3{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        let text_res = args.get(2);
+        if space_res.is_number() && fpath_res.is_string() && text_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             let text = format!("{}", text_res.as_string().unwrap().to_string().unwrap());
             let res = FileIO::write_text(fpath, text.as_str());
             return Ok(Value::Boolean(res));
         }else{
-            error!("Invalid argument, expected string arguments");
+            error!("Invalid argument, expected number, string, string arguments");
             return Ok(Value::Boolean(false));    
         }
     }else{
-        error!("Invalid argument, expected 2 arguments");
+        error!("Invalid argument, expected 3 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -88,24 +104,22 @@ pub fn file_write_text(inv: Invocation) -> Result<Value, DuccError> {
 pub fn file_append_text(inv: Invocation) -> Result<Value, DuccError> {
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 2{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        let text_res = args.get(1);
-        
-        if fpath_res.is_string() && text_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 3{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        let text_res = args.get(2);
+        if space_res.is_number() && fpath_res.is_string() && text_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             let text = format!("{}", text_res.as_string().unwrap().to_string().unwrap());
             let res = FileIO::append_text(fpath, text.as_str());
             return Ok(Value::Boolean(res));
         }else{
-            error!("Invalid argument, expected string arguments");
+            error!("Invalid argument, expected number, string, string arguments");
             return Ok(Value::Boolean(false));    
         }
     }else{
-        error!("Invalid argument, expected 2 arguments");
+        error!("Invalid argument, expected 3 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -113,13 +127,12 @@ pub fn file_append_text(inv: Invocation) -> Result<Value, DuccError> {
 pub fn file_read_text(inv: Invocation) -> Result<Value, DuccError> {
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             let mut text = String::from("");
             let res = FileIO::read_text(fpath.clone(), &mut text);
             if res{
@@ -129,11 +142,11 @@ pub fn file_read_text(inv: Invocation) -> Result<Value, DuccError> {
                 return Ok(Value::Null);
             }
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Null);
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 argument");
         return Ok(Value::Null);
     }
 }
@@ -141,13 +154,12 @@ pub fn file_read_text(inv: Invocation) -> Result<Value, DuccError> {
 pub fn file_read(inv: Invocation) -> Result<Value, DuccError> {
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             let mut buf = Vec::<u8>::new();
             let res = FileIO::read_bytes(fpath.clone(), &mut buf);
             if res{
@@ -157,11 +169,11 @@ pub fn file_read(inv: Invocation) -> Result<Value, DuccError> {
                 return Ok(Value::Null);
             }
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Null);    
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 argument");
         return Ok(Value::Null);
     }
 }
@@ -169,15 +181,13 @@ pub fn file_read(inv: Invocation) -> Result<Value, DuccError> {
 pub fn file_write(inv: Invocation) -> Result<Value, DuccError> {
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 2{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-        
-        let fpath_res = args.get(0);
-        let bytes_res = args.get(1);
-
-        if fpath_res.is_string() && bytes_res.is_bytes(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 3{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        let bytes_res = args.get(2);
+        if space_res.is_number() && fpath_res.is_string() && bytes_res.is_bytes(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             let buf = bytes_res.as_bytes().unwrap().to_vec();
             let res = FileIO::write_bytes(fpath, &buf);
             return Ok(Value::Boolean(res));
@@ -194,20 +204,19 @@ pub fn file_write(inv: Invocation) -> Result<Value, DuccError> {
 pub fn file_remove(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FileIO::remove(fpath)));
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Boolean(false));
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid arguments, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -215,20 +224,19 @@ pub fn file_remove(inv: Invocation) -> Result<Value, DuccError>{
 pub fn folder_create(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FolderIO::create(fpath)));
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Boolean(false));
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -236,20 +244,19 @@ pub fn folder_create(inv: Invocation) -> Result<Value, DuccError>{
 pub fn folder_create_all(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FolderIO::create_all(fpath)));
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Boolean(false));
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -257,20 +264,19 @@ pub fn folder_create_all(inv: Invocation) -> Result<Value, DuccError>{
 pub fn folder_remove(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FolderIO::remove(fpath)));
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Boolean(false));
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
@@ -278,20 +284,19 @@ pub fn folder_remove(inv: Invocation) -> Result<Value, DuccError>{
 pub fn folder_remove_all(inv: Invocation) -> Result<Value, DuccError>{
     let engine = inv.ducc;
     let args = inv.args;
-    if args.len() == 1{
-        let robj:Object = engine.globals().get(GALE_KEY).unwrap();
-        let data_root:String = robj.get(DATA_ROOT_KEY).unwrap();
-
-        let fpath_res = args.get(0);
-        if fpath_res.is_string(){
-            let fpath = format!("{}/{}", data_root, fpath_res.as_string().unwrap().to_string().unwrap());
+    if args.len() == 2{
+        let space_res = args.get(0);
+        let fpath_res = args.get(1);
+        if space_res.is_number() && fpath_res.is_string(){
+            let base_dir = get_base_dir(engine, space_res.as_number().unwrap() as u8);
+            let fpath = format!("{}/{}", base_dir, fpath_res.as_string().unwrap().to_string().unwrap());
             return Ok(Value::Boolean(FolderIO::remove_all(fpath)));
         }else{
-            error!("Invalid argument, expected string argument");
+            error!("Invalid argument, expected number, string argument");
             return Ok(Value::Boolean(false));
         }
     }else{
-        error!("Invalid argument, expected 1 argument");
+        error!("Invalid argument, expected 2 arguments");
         return Ok(Value::Boolean(false));
     }
 }
